--- DQM/Physics/.gitignore
@@ -1,2 +1,0 @@
-*.root
-*.log
